+++ main/app/modules/takeout/domain/model/takeout_settings.go
@@ -9,8 +9,8 @@ type TakeoutSettings struct {
 	IsEnabled int `gorm:"column:is_enabled" json:"is_enabled"`
 
 	// 自动接单配置
-	AutoAccept int   `gorm:"column:auto_accept" json:"auto_accept"`
-	MaxAmount  int64 `gorm:"column:max_amount" json:"max_amount"`
+	AutoAccept int     `gorm:"column:auto_accept" json:"auto_accept"`
+	MaxAmount  float64 `gorm:"column:max_amount" json:"max_amount"`
 
 	// 平台特定配置（JSON 格式）
 	PlatformConfig string `gorm:"column:platform_config;type:text" json:"platform_config"`
